Remove hacky local_fields craziness.
This was originally added to fix #13, but I can't reproduce the issue reported in #13 now (in django 1.5.x at least.)

It breaks compatibility with django 1.7, so if someone comes across bad sequence values again we can find a different, less hacky fix for it.

afaik this is the last remaining issue for django 1.7 compatibility. fixes #16. please reopen if you find more 1.7 issues.