BACKWARD-INCOMPATIBLE. Don't auto-cast if `type` is deferred
If you use QuerySet.defer() or QuerySet.only() to defer fields,
and one of the fields you defer is the `type` field, this change
means that typedmodels will no longer automatically cast the
model instances in the queryset to the correct type.

Previously, we did automatically cast the instances, at the expense
of doing an extra query *per instance* (that's just how field deferral
works.)

With Django 3.2 (due to some complicated internal changes in the ORM)
that no longer works.